Make the following changes:
,

Find this element:

  <ChatMessageDisplay message={"id":"msg-1747729545321-sjl1q1d3x","role":"user","content":"Triggered action: generateDesignPrompts","timestamp":1747729545321,"attachedFiles":[],"canRegenerate":false}>

with the text content:

    02:25 PMTriggered action: generateDesignPrompts

Relevant files:

    - https://6000-firebase-studio-1747397889806.cluster-ikxjzjhlifcwuroomfkjrx437g.cloudworkstations.dev/_next/static/chunks/node_modules_dc3972f1._.js
    - https://6000-firebase-studio-1747397889806.cluster-ikxjzjhlifcwuroomfkjrx437g.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747397889806.cluster-ikxjzjhlifcwuroomfkjrx437g.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

I want, when the user clicks on the regenerate button, normally now he can't edit his input/prompt. So it generate Results from the same prompt, so I want the user to be able to edit his input/prompt and regenerate from here, so at the region button in the user input section.